fix: 修复问题单2权限校验与输入验证17个问题（Commit C2）
- H-2: Socket数据通道port_字段增加归属校验
- M-2: SSAP Server 4个方法增加appId归属校验
- M-3: SSAP property/method permission_写读宽度对齐uint16_t
- M-8: SSAP onCallMethod回调判空
- M-9: SM配对确认codeAlgoCap逐字节比对
- M-15: 电池观察者回调增加权限校验与地址脱敏
- 批05: central_manager_proxy callback判空
- 批04: asc_proxy计数校验与对端对齐
- 批26-27: OnPairConfirmed权限提升为MANAGE_NEARLINK
- L-3: SLE_START_ADVERTISING权限对齐MANAGE_NEARLINK
- L-4: GetServicesByUuid补齐数量前缀
- L-6: hadm Deregister校验callback与hadmId绑定
- L-10: devd frameType校验前置
- 批05: RemoveAllPairs WriteBool改WriteInt32
- 批05: hid_host_stub字段数与客户端对齐

+++ ipc_parcel/parcel/nearlink_ssap_property_parcel.cpp
@@ -53,7 +53,7 @@ bool NearlinkSsapPropertyParcel::Marshalling(Parcel &parcel) const
             return false;
         }
     }
-    if (!parcel.WriteInt32(permission_)) {
+    if (!parcel.WriteUint16(permission_)) {
         return false;
     }
     return true;
